extracted mithril comp: doc, pbxpages, persontyle
