+++ tomatic/ui/src/mithril/graella.js
@@ -5,26 +5,25 @@ m.prop = require('mithril/stream')
 
 var Snackbar = require('polythene-mithril-snackbar').Snackbar
 var Dialog = require('polythene-mithril-dialog').Dialog
-var Card = require('polythene-mithril-card').Card
 var Tabs = require('polythene-mithril-tabs').Tabs
 
 var iconMenu = require('mmsvg/google/msvg/navigation/menu')
 var iconMore = require('mmsvg/google/msvg/navigation/more-vert')
 
 var Tomatic = require('../services/tomatic')
-var QueueMonitor = require('./components/queuemonitor')
-var luminance = require('./components/colorutils').luminance
-var contrast = require('./components/colorutils').contrast
 var Login = require('./components/login')
 var Persons = require('./components/persons')
 var TimeTablePage = require('./components/timetablepage')
 var MenuButton = require('./components/menubutton')
 var LoginPage = require('./components/loginpage')
+var PersonStyles = require('./components/personstyles').default
+var Doc = require('./components/doc').default
 
 var css = require('polythene-css')
 var customStyle = require('./style.styl')
 
 var CallInfoPage = require('./components/callinfopage')
+var PbxPage = require('./components/pbxpage')
 
 css.addLayoutStyles()
 css.addTypography()
@@ -48,32 +47,6 @@ const SnackbarLogger = function () {
 	}
 }
 Tomatic.loggers.push(SnackbarLogger())
-
-var Todo = function (message) {
-	return m(Card, {
-		content: [
-			{
-				primary: {
-					title: 'TODO',
-					subtitle: message,
-				},
-			},
-		],
-	})
-}
-
-var Doc = function (message) {
-	return m(Card, {
-		content: [
-			{
-				primary: {
-					title: 'Info',
-					subtitle: message,
-				},
-			},
-		],
-	})
-}
 
 const scriptLauncherBase = 'http://tomatic.somenergia.lan:5000'
 const menuOptions = function () {
@@ -168,22 +141,6 @@ const toolbarRow = function (title) {
 	])
 }
 
-var PbxPage = {
-	view: function () {
-		return m('', [
-			Doc(
-				m(
-					'',
-					'Visualitza les línies que estan actualment rebent trucades. ' +
-						"Feu click al damunt per pausar-les o al signe '+' per afegir-ne"
-				)
-			),
-			m('h2[style=text-align:center]', 'Linies en cua'),
-			m(QueueMonitor),
-		])
-	},
-}
-
 var PersonsPage = {
 	view: function () {
 		return m('', [
@@ -194,52 +151,6 @@ var PersonsPage = {
 			Persons(Tomatic.persons().extensions),
 		])
 	},
-}
-
-var PersonStyles = function () {
-	var persons = Tomatic.persons()
-	return m(
-		'style',
-		Object.keys(persons.colors || {}).map(function (name) {
-			var color = '#' + persons.colors[name]
-			var darker = '#' + luminance(color, -0.3)
-			var linecolor = contrast(persons.colors[name])
-			return (
-				'.' +
-				name +
-				', .graella .' +
-				name +
-				' {\n' +
-				'  background-color: ' +
-				color +
-				';\n' +
-				'  border-color: ' +
-				darker +
-				';\n' +
-				'  border-width: 20pt;\n' +
-				'  color: ' +
-				linecolor +
-				';\n' +
-				'}\n' +
-				'.pe-dark-theme .' +
-				name +
-				', .pe-dark-theme .graella .' +
-				name +
-				' {\n' +
-				'  background-color: ' +
-				darker +
-				';\n' +
-				'  border-color: ' +
-				color +
-				';\n' +
-				'  border-width: 2pt;\n' +
-				'  color: ' +
-				linecolor +
-				';\n' +
-				'}\n'
-			)
-		})
-	)
 }
 
 const applicationPages = ['Graelles', 'Centraleta', 'Persones', 'Trucada']
